+++ src/services/landlord/application/Decision/types.ts
@@ -1,9 +1,7 @@
-import { IApplication } from '../../../../models/Application'
-
 export type ApproveForm = {
-  application: IApplication
+  applicationId: string
 }
 
 export type DeclineForm = {
-  application: IApplication
+  applicationId: string
 }